+++ src/hooks/index.js
@@ -6,3 +6,4 @@ export { default as useFetchMonthlyDiaryData } from './useFetchMonthlyDiaryData.
 export { default as useFetchAllBuddyData } from './useFetchAllBuddyData.jsx';
 export { default as useCheckAvailability } from './useCheckAvailability.jsx';
 export { default as useBlocker } from './useBlocker.jsx';
+export { default as useDiaryActions } from './useDiaryActions';
